+++ plugin/src/Assets.lua
@@ -24,6 +24,7 @@ local Assets = {
 			Close = "rbxassetid://6012985953",
 			Back = "rbxassetid://6017213752",
 			Reset = "rbxassetid://10142422327",
+			Expand = "rbxassetid://12045401097",
 		},
 		Diff = {
 			Add = "rbxassetid://10434145835",