+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Add `plugin` flag to the `build` command that outputs to the local plugins folder ([#735])
 * Added better support for `Font` properties ([#731])
 * Add new plugin template to the `init` command ([#738])
+* Added rich Source diffs in patch visualizer ([#748])
 
 [#745]: https://github.com/rojo-rbx/rojo/pull/745
 [#668]: https://github.com/rojo-rbx/rojo/pull/668
@@ -42,6 +43,7 @@
 [#735]: https://github.com/rojo-rbx/rojo/pull/735
 [#731]: https://github.com/rojo-rbx/rojo/pull/731
 [#738]: https://github.com/rojo-rbx/rojo/pull/738
+[#748]: https://github.com/rojo-rbx/rojo/pull/748
 
 ## [7.3.0] - April 22, 2023
 * Added `$attributes` to project format. ([#574])